♻️ refactor: 댓글 코드 수정

+++ src/main/java/com/example/Backend/domain/repository/PostRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface PostRepository extends JpaRepository<Post, Integer>, PostQueryDsl {
+public interface PostRepository extends JpaRepository<Post, Long>, PostQueryDsl {
     Optional<Post> findPostById(Long id);
 }
